// src/lib/types/enums.ts
export const enum Prize {
	First = '1st Place',
	Second = '2nd Place',
	Third = '3rd Place',
	Fourth = '4th Place',
	Fifth = '5th Place',
	HM = 'Honorable Mention'
}

export const enum Tier {
	Platinum = 'Platinum Tier',
	Gold = 'Gold Tier',
	Silver = 'Silver Tier',
	Bronze = 'Bronze Tier',
	Alum = 'Alum Sponsors',
	Null = ''
}